requirements.txt
fastapi>=0.68.0
uvicorn>=0.15.0
python-multipart>=0.0.5
websockets>=10.0
langchain>=0.1.0
langchain-community>=0.1.0
langchain-core>=0.1.0
langchain-google-genai>=0.1.0
langchain-groq>=0.1.0
langchain-huggingface>=0.1.0
langgraph>=0.1.0
tavily-python>=0.7.0
python-dotenv>=0.19.0
pydantic>=1.8.0
typing-extensions>=3.10.0
chromadb>=0.4.0
pypdf>=3.8.0
sentence-transformers>=2.2.0
tiktoken>=0.4.0
requests>=2.28.0